more animation work, with parser also moved all object classes
