+++ make_video.sh
@@ -6,6 +6,7 @@ if [[ -z $1 ]]; then
   echo "Supply result directory!"
   exit
 fi
+
 if [[ -z $2 ]]; then 
   framerate=30
 else
@@ -13,5 +14,5 @@ else
 fi
 
 pushd "$1"
-ffmpeg -framerate $framerate -i "result_%05d.png" -vcodec libx264 -pix_fmt yuv420p -y output.mp4
+ffmpeg -framerate $framerate -i "result_%05d.png" -vcodec libx264 -pix_fmt yuv420p -y video.mp4
 popd
